나침반 방향 , 현재높이 단위 수정

+++ danusys-drone/src/main/java/com/danusys/web/drone/dto/response/Gps.java
@@ -11,10 +11,10 @@ import lombok.Setter;
 public class Gps {
     private Double gpsX;  //z
     private Double gpsY;  //y
-    private int currentHeight; // 높이
+    private Double currentHeight; // 높이
     private int wpDist; // 목표까지의거리
     private String missionType; //미션 종류 (귀환,이륙,이동 등등)
-    private int heading; //방향 0이 north
+    private Double heading; //방향 0이 north
   //  private float airspeed; //수직속도
  //   private float groundspeed;//수평속도
